+++ src/dask_awkward/utils.py
@@ -3,9 +3,6 @@ from __future__ import annotations
 from collections.abc import Callable, Mapping
 from typing import Any
 
-import awkward._v2 as ak
-import numpy as np
-
 
 class DaskAwkwardNotImplemented(NotImplementedError):
     NOT_SUPPORTED_MSG = """
@@ -75,19 +72,6 @@ def borrow_docstring(original: Callable) -> Callable:
     return wrapper
 
 
-def empty_typetracer() -> ak.Array:
-    """Instantiate a typetracer array with unknown length.
-
-    Returns
-    -------
-    ak.Array
-        Length-less typetracer array (content-less array).
-
-    """
-    a = ak.Array([])
-    return ak.Array(a.layout.typetracer.forget_length())
-
-
 def hyphenize(x: str) -> str:
     """Replace underscores with hyphens.
 
@@ -133,43 +117,3 @@ def is_empty_slice(s: Any) -> bool:
     if s.step is not None:
         return False
     return True
-
-
-def normalize_single_outer_inner_index(
-    divisions: tuple[int, ...], index: int
-) -> tuple[int, int]:
-    """Determine partition index and inner index for some divisions.
-
-    Parameters
-    ----------
-    divisions : tuple[int, ...]
-        The divisions of a Dask awkward collection.
-    index : int
-        The overall index (for the complete collection).
-
-    Returns
-    -------
-    partition_index : int
-        Which partition in the collection.
-    new_index : int
-        Which inner index in the determined partition.
-
-    Examples
-    --------
-    >>> from dask_awkward.utils import normalize_single_outer_inner_index
-    >>> divisions = (0, 3, 6, 9)
-    >>> normalize_single_outer_inner_index(divisions, 0)
-    (0, 0)
-    >>> normalize_single_outer_inner_index(divisions, 5)
-    (1, 2)
-    >>> normalize_single_outer_inner_index(divisions, 8)
-    (2, 2)
-
-    """
-    if index < 0:
-        index = divisions[-1] + index
-    if len(divisions) == 2:
-        return (0, index)
-    partition_index = int(np.digitize(index, divisions)) - 1
-    new_index = index - divisions[partition_index]
-    return (partition_index, new_index)
