+++ tests/test_behavior.py
@@ -7,7 +7,7 @@ from awkward._v2.behaviors.mixins import mixin_class as ak_mixin_class
 from awkward._v2.behaviors.mixins import mixin_class_method as ak_mixin_class_method
 
 import dask_awkward as dak
-from dask_awkward.testutils import assert_eq
+from dask_awkward.lib.testutils import assert_eq
 
 behaviors: dict = {}
